crawldelay seconds->nanoseconds

+++ types/domain.go
@@ -1,9 +1,10 @@
 package types
 
+//times are in nanoseconds
 type Domain struct {
 	Name        string
-	CrawlDelay  int
-	LastCrawled int
+	CrawlDelay  int64
+	LastCrawled int64
 	Allowed     []string
 	Disallowed  []string
 }
